requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.0.0
requests==2.31.0
gunicorn==20.1.0